requirements.txt
torch
torchvision
pandas
matplotlib
scikit-learn
imagehash
pillow
